Script factor generation for calc. compensation of Tor exit operators.
From JSON data taken from Compass, it creates a multidimensional array of
country codes and the probability of exiting from that country. Then, it
generates a windorized standard deviation and a trimmed standard deviation of
the probabilities within that array. Next, it takes the standard deviation of
all combined exit-by-country probabilities, subtracts either the winsorized or
trimmed standard deviation of all combined exit-by-country probabilities, adds
the probability of for exiting in that country, takes the absolute value of
this whole mess and computes the inverse squared: This gives us an
incentivization factor for disbursal of funds to exit relay operators in
countries with less exit relays.