Correctly check commit type when rev_cur is zero
On creation of a ref, the script is called with rev_old=0, so on the first
iteration of the loop $rev_cur is 0. Attempting to run "git rev-list" on that
gives an error, so $merge will be empty. Then, the attempt to run
"git cat-file -t" will give an error, leading to an empty $commit_type.

Rearrange the tests so we only run "git rev-list" and "git cat-file" on non-
zero ID:s, and assume a type of "commit" when the previous commit ID was 0.