Be more paranoid about grep:ing GnuPG output
By default, Git commands that verify signatures output not only the result
of the verification, but also the entire message. Since we're checking the
verification result with Grep, this means we can be fooled by strings in
the commit messages.

Make sure to throw away commit messages before checking result strings.