Properly check for commits and merges on master
The old code considered all commits between old and new HEAD when checking
whether there would be any normal commits on master, or only merges. This is
wrong, since that list of commits includes every commit on the feeature branch
you are trying to merge. Instead, check that the new HEAD points to a merge
commit that has the old HEAD as one of its parents.