Properly check tags for annotations and signatures
When a tag is created, the hook script gets called with an old ref of zero and
a new ref of the commit pointed to by the tag. However, this commit is already
in the commit graph (since it is on a branch), so our "git rev-list" command
to find all commits affected by this push will return an empty list.

Use this empty list to signal that we're dealing with tags, and handle them
outside of the main loop.